crawler-descryptor/requirements.txt
httpx>=0.27
pycryptodome>=3.20
frida-tools>=12.0
